LayoutTests/fast/shadow-dom/media-shadow-manipulation.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<script>
window.jsTestIsAsync = true;

function runTest() {
    shouldBe('bounds = video_element.getBoundingClientRect(); document.caretRangeFromPoint(bounds.left + 10, bounds.bottom - 10)', 'null');
    finishJSTest();
}
</script>
</head>
<body onload="runTest()">
    <video id="video_element" style="position:absolute; height: 240px; width: 320px;" controls></video>
</body>
</html>
