LayoutTests/fast/shadow-dom/media-shadow-manipulation-expected.txt

PASS bounds = video_element.getBoundingClientRect(); document.caretRangeFromPoint(bounds.left + 10, bounds.bottom - 10) is null
PASS successfullyParsed is true

TEST COMPLETE
